feat: add path parameter to find() function
The update consists of refactoring the code in the find() function, then adding a path parameter, and adding a log program to the exception_factory() function by taking the value of the message argument.

Using the tree that belongs to the rich module for output to the find() function is so difficult, there needs to be a new code update but has the same logical path. This will be the next task